Backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.10.1
xgboost==1.7.6
shap==0.42.1
numpy==1.24.3
joblib==1.3.1
python-dotenv==1.0.0
utils.explainer